feat!(editor): refactor core and extra editor plugins
