+++ lua/config/editor/whichkey.lua
@@ -84,60 +84,10 @@ local mappings = {
       ["r"] = { "<cmd>:source $MYVIMRC<CR>", "Reload" },
     },
 
-    ["b"] = {
-      d = { "<cmd>Bdelete!<CR>", "Close Buffer" },
-      l = { "<cmd>lua require('telescope.builtin').buffers()<cr>", "Buffer list" },
-      s = { "<cmd>lua require('telescope.builtin').current_buffer_fuzzy_find()<cr>", "Buffer Search" },
-      S = { "<cmd>lua require('telescope.builtin').treesitter()<cr>", "Buffer Symbols" },
-      -- Close all other buffers, performing the following operations:
-      -- * :w - Save the current file
-      -- * %bdelete - Close all buffers
-      -- * edit# - open the last edited file (the buffer we want to keep)
-      -- * bdelete# - Close the unnamed buffer
-      O = { '<cmd>:w <bar> %bdelete <bar> edit# <bar> bdelete# <bar> normal `"<CR>', "Delete other buffers" },
-    },
-
-    ["f"] = {
-      name = "file",
-      f = {
-        "<cmd>lua require('telescope.builtin').find_files({ hidden = true, search_dirs = { '/media/procore', '~/github.com', '~/.config/nvim', '/etc/dotfiles' }})<cr>",
-        "Find files",
-      },
-      s = { "<cmd>silent w!<CR>", "Save" },
-      S = {
-        "<cmd>lua vim.lsp.buf.format()<CR><cmd>silent w!<CR>",
-        "Format and Save",
-      },
-    },
-
     ["o"] = {
       name = "open",
 
-      p = {
-        function() require("neo-tree.command").execute({ toggle = true, dir = vim.uv.cwd(), position = "left" }) end,
-        "Explorer",
-      },
-      P = {
-        function() require("neo-tree.command").execute({ toggle = true, dir = vim.uv.cwd(), position = "float" }) end,
-        "Explorer Float",
-      },
       t = { "<cmd>ToggleTerm<cr>", "Terminal" },
-    },
-
-    ["h"] = {
-      name = "help",
-      ["'"] = { "<cmd>Inspect<CR>", "Inspect Element" },
-      a = { "<cmd>lua require('telescope.builtin').autocommands()<CR>", "Display Autocommands" },
-      h = { "<cmd>lua require('telescope.builtin').highlights()<CR>", "Display Highlights" },
-      c = { "<cmd>lua require('telescope.builtin').commands()<CR>", "Display Commands" },
-      C = { "<cmd>lua require('telescope.builtin').colorscheme()<CR>", "Display Colorschemes" },
-      f = { "<cmd>lua require('telescope.builtin').filetypes()<CR>", "Display Filetypes" },
-      k = { "<cmd>lua require('telescope.builtin').keymaps()<CR>", "Display Keymaps" },
-      n = { "<cmd>Notifications<CR>", "Display Notifications" },
-      -- m = { "<cmd>messages<CR>", "Display Messages" },
-      m = { "<cmd>Noice telescope<CR>", "Display Messages" },
-      o = { "<cmd>lua require('telescope.builtin').vim_options()<CR>", "Display Options" },
-      v = { "<cmd>lua require('telescope.builtin').treesitter()<CR>", "Display Buffer Variables" },
     },
 
     -- Window Management
@@ -155,15 +105,6 @@ local mappings = {
 
       -- Window clean up
       d = { "<C-w>c<CR>", "Delete Window" },
-    },
-
-    ["s"] = {
-      name = "search",
-      p = {
-        function() require("config.editor.telescope.custom_pickers").live_grep() end,
-        "Find Text",
-      },
-      i = { "<cmd>IconPickerInsert<cr>", "Find Icon" },
     },
 
     -- Buffer Movement
@@ -233,75 +174,10 @@ local mappings = {
       },
     },
   },
-
-  ["f"] = {
-    ["d"] = { "zd", "Delete fold under cursor" },
-    ["o"] = {
-      "zo<cmd>IndentBlanklineEnable<CR>",
-      "Open fold under cursor",
-    },
-    ["O"] = { "zO", "Open all folds under cursor" },
-    ["c"] = { "zc", "Close fold under cursor" },
-    ["C"] = { "zC", "Close all folds under cursor" },
-    ["a"] = {
-      "za<cmd>IndentBlanklineEnable<CR>",
-      "Toggle fold under cursor",
-    },
-    ["A"] = { "zA", "Toggle all folds under cursor" },
-    ["v"] = { "zv", "Show cursor line" },
-    ["M"] = { require("ufo").closeAllFolds, "Close all folds" },
-    ["R"] = { require("ufo").openAllFolds, "Open all folds" },
-    ["m"] = { "zm", "Fold more" },
-    ["r"] = { "zr", "Fold less" },
-    ["x"] = { "zx", "Update folds" },
-    ["z"] = { "zz", "Center this line" },
-    ["t"] = { "zt", "Top this line" },
-    ["b"] = { "zb", "Bottom this line" },
-    ["g"] = { "zg", "Add word to spell list" },
-    ["w"] = { "zw", "Mark word as bad/misspelling" },
-    ["e"] = { "ze", "Right this line" },
-    ["E"] = { "zE", "Delete all folds in current buffer" },
-    ["s"] = { "zs", "Left this line" },
-    ["H"] = { "zH", "Half screen to the left" },
-    ["L"] = { "zL", "Half screen to the right" },
-  },
-  ["s"] = {
-    ["b"] = { "<cmd>Telescope git_branches<cr>", "Checkout branch" },
-    ["c"] = { "<cmd>Telescope colorscheme<cr>", "Colorscheme" },
-    ["h"] = { "<cmd>Telescope help_tags<cr>", "Find Help" },
-    ["M"] = { "<cmd>Telescope man_pages<cr>", "Man Pages" },
-    ["r"] = { "<cmd>Telescope oldfiles<cr>", "Open Recent File" },
-    ["R"] = { "<cmd>Telescope registers<cr>", "Registers" },
-    ["k"] = { "<cmd>Telescope keymaps<cr>", "Keymaps" },
-    ["C"] = { "<cmd>Telescope commands<cr>", "Commands" },
-  },
-  --  ["g"] = {
-  --    ["d"] = { "<cmd>Telescope lsp_definitions<cr>", "Go to definition" },
-  --    -- ["d"] = { "<cmd>lua vim.lsp.buf.definition()<cr>", "Go to definition" },
-  --    ["r"] = { "<cmd>Telescope lsp_references<cr>", "Go to references" },
-  --    ["i"] = {
-  --      "<cmd>Telescope lsp_implementations<cr>",
-  --      "Go to implementations",
-  --    },
-  --    ["b"] = { "<cmd>BufferLinePick<CR>", "Bufferline: pick buffer" },
-  --  },
-  ["<C-=>"] = {
-    function() vim.g.neovide_scale_factor = vim.g.neovide_scale_factor * 1.05 end,
-    "Increase Font",
-  },
-  ["<C-->"] = {
-    function() vim.g.neovide_scale_factor = vim.g.neovide_scale_factor * (1 / 1.05) end,
-    "Decrease Font",
-  },
 }
 
 which_key.register(mappings, { mode = "n", prefix = "" })
 which_key.register({
-  ["c"] = {
-    ["e"] = {
-      require("util").runlua,
-      "Run Lua",
-    },
-  },
+  ["c"] = { ["e"] = { require("util").runlua, "Run Lua" } },
 }, { mode = "v", prefix = "<leader>" })
 which_key.register({ ["f"] = { "zf", "Create fold" } }, { mode = "v", prefix = "f" })
