+++ lua/plugins/editor.lua
@@ -1,3 +1,5 @@
+local icons = require("core.icons")
+
 return {
   -- File Navigation
   {
@@ -6,23 +8,199 @@ return {
       "nvim-tree/nvim-web-devicons",
     },
     cmd = "Neotree",
-    deactivate = function()
-      vim.cmd([[Neotree close]])
+    keys = {
+      {
+        "<leader>fe",
+        function() require("neo-tree.command").execute({ toggle = true, dir = CoreUtil.root() }) end,
+        desc = "Explorer NeoTree (Root Dir)",
+      },
+      {
+        "<leader>fE",
+        function() require("neo-tree.command").execute({ toggle = true, dir = vim.uv.cwd() }) end,
+        desc = "Explorer NeoTree (cwd)",
+      },
+      { "<leader>e", "<leader>fe", desc = "Explorer NeoTree (Root Dir)", remap = true },
+      { "<leader>E", "<leader>fE", desc = "Explorer NeoTree (cwd)",      remap = true },
+      {
+        "<leader>ge",
+        function() require("neo-tree.command").execute({ source = "git_status", toggle = true }) end,
+        desc = "Git Explorer",
+      },
+      {
+        "<leader>be",
+        function() require("neo-tree.command").execute({ source = "buffers", toggle = true }) end,
+        desc = "Buffer Explorer",
+      },
+    },
+    deactivate = function() vim.cmd([[Neotree close]]) end,
+    init = function()
+      -- FIX: use `autocmd` for lazy-loading neo-tree instead of directly requiring it,
+      -- because `cwd` is not set up properly.
+      vim.api.nvim_create_autocmd("BufEnter", {
+        group = vim.api.nvim_create_augroup("Neotree_start_directory", { clear = true }),
+        desc = "Start Neo-tree with directory",
+        once = true,
+        callback = function()
+          if package.loaded["neo-tree"] then
+            return
+          else
+            local stats = vim.uv.fs_stat(vim.fn.argv(0))
+            if stats and stats.type == "directory" then require("neo-tree") end
+          end
+        end,
+      })
     end,
-    init = require("config.editor.neo-tree").init,
-    opts = require("config.editor.neo-tree").opts,
-    config = require("config.editor.neo-tree").config,
+    opts = {
+      close_if_last_window = true, -- Close Neo-tree if it is the last window left in the tab
+      source_selector = {
+        winbar = true,             -- toggle to show selector on winbar
+        content_layout = "center",
+        tabs_layout = "equal",
+        show_separator_on_edge = true,
+        sources = {
+          { source = "filesystem", display_name = "󰉓" },
+          { source = "buffers", display_name = "󰈙" },
+          { source = "git_status", display_name = "" },
+          -- { source = "document_symbols", display_name = "o" },
+          { source = "diagnostics", display_name = "󰒡" },
+        },
+      },
+      default_component_configs = {
+        indent = {
+          with_expanders = true, -- if nil and file nesting is enabled, will enable expanders
+          expander_collapsed = "",
+          expander_expanded = "",
+          expander_highlight = "NeoTreeExpander",
+        },
+        modified = { symbol = icons.git.modified },
+        git_status = { symbols = icons.git },
+        diagnostics = { symbols = icons.diagnostics },
+      },
+      window = {
+        mappings = {
+          ["l"] = "open",
+          ["h"] = "close_node",
+          ["<space>"] = "none",
+          ["Y"] = {
+            function(state)
+              local node = state.tree:get_node()
+              local path = node:get_id()
+              vim.fn.setreg("+", path, "c")
+            end,
+            desc = "Copy Path to Clipboard",
+          },
+          ["O"] = {
+            function(state) require("lazy.util").open(state.tree:get_node().path, { system = true }) end,
+            desc = "Open with System Application",
+          },
+          ["P"] = { "toggle_preview", config = { use_float = false } },
+        },
+      },
+      filesystem = {
+        bind_to_cwd = false,
+        filtered_items = {
+          visible = false, -- when true, they will just be displayed differently than normal items
+          hide_dotfiles = false,
+          hide_gitignored = false,
+        },
+        follow_current_file = {
+          enabled = true,
+        },
+        -- This will find and focus the file in the active buffer every
+        -- time the current file is changed while the tree is open.
+        group_empty_dirs = false,               -- when true, empty folders will be grouped together
+        hijack_netrw_behavior = "open_default", -- netrw disabled, opening a directory opens neo-tree
+        -- in whatever position is specified in window.position
+        -- "open_current",  -- netrw disabled, opening a directory opens within the
+        -- window like netrw would, regardless of window.position
+        -- "disabled",    -- netrw left alone, neo-tree does not handle opening dirs
+        use_libuv_file_watcher = true, -- This will use the OS level file watchers to detect changes
+        -- instead of relying on nvim autocmd events.
+      },
+      buffers = {
+        show_unloaded = true,
+      },
+      git_status = {
+        window = {
+          position = "left",
+        },
+      },
+    },
+    config = function(_, opts)
+      require("neo-tree").setup(opts)
+      vim.api.nvim_create_autocmd("TermClose", {
+        pattern = "*lazygit",
+        callback = function()
+          if package.loaded["neo-tree.sources.git_status"] then require("neo-tree.sources.git_status").refresh() end
+        end,
+      })
+    end,
   },
 
-  -- Fuzzy Search
+  -- search/replace in multiple files
   {
-    "nvim-telescope/telescope.nvim",
-    cmd = "Telescope",
-    version = false,
-    dependencies = {
-      "JordanFaust/project.nvim",
+    "nvim-pack/nvim-spectre",
+    build = false,
+    cmd = "Spectre",
+    opts = { open_cmd = "noswapfile vnew" },
+    -- stylua: ignore
+    keys = {
+      { "<leader>sr", function() require("spectre").open() end, desc = "Replace in Files (Spectre)" },
     },
-    config = function() require("config.editor.telescope") end,
+  },
+
+  -- Flash enhances the built-in search functionality by showing labels
+  -- at the end of each match, letting you quickly jump to a specific
+  -- location.
+  -- {
+  --   "folke/flash.nvim",
+  --   event = "VeryLazy",
+  --   vscode = true,
+  --   ---@type Flash.Config
+  --   opts = {},
+  --   -- stylua: ignore
+  --   keys = {
+  --     { "s",     mode = { "n", "x", "o" }, function() require("flash").jump() end,              desc = "Flash" },
+  --     { "S",     mode = { "n", "o", "x" }, function() require("flash").treesitter() end,        desc = "Flash Treesitter" },
+  --     { "r",     mode = "o",               function() require("flash").remote() end,            desc = "Remote Flash" },
+  --     { "R",     mode = { "o", "x" },      function() require("flash").treesitter_search() end, desc = "Treesitter Search" },
+  --     { "<c-s>", mode = { "c" },           function() require("flash").toggle() end,            desc = "Toggle Flash Search" },
+  --   },
+  -- },
+
+  -- which-key helps you remember key bindings by showing a popup
+  -- with the active keybindings of the command you started typing.
+  {
+    "folke/which-key.nvim",
+    event = "VeryLazy",
+    opts = {
+      plugins = { spelling = true },
+      defaults = {
+        mode = { "n", "v" },
+        ["g"] = { name = "+goto" },
+        ["gs"] = { name = "+surround" },
+        ["z"] = { name = "+fold" },
+        ["]"] = { name = "+next" },
+        ["["] = { name = "+prev" },
+        ["<leader><tab>"] = { name = "+tabs" },
+        ["<leader>b"] = { name = "+buffer" },
+        ["<leader>c"] = { name = "+code" },
+        ["<leader>f"] = { name = "+file/find" },
+        ["<leader>g"] = { name = "+git" },
+        ["<leader>gh"] = { name = "+hunks", ["_"] = "which_key_ignore" },
+        ["<leader>h"] = { name = "+help" },
+        ["<leader>q"] = { name = "+quit/session" },
+        ["<leader>s"] = { name = "+search" },
+        ["<leader>u"] = { name = "+ui" },
+        ["<leader>w"] = { name = "+windows" },
+        ["<leader>x"] = { name = "+diagnostics/quickfix" },
+      },
+    },
+    config = function(_, opts)
+      local wk = require("which-key")
+      wk.setup(opts)
+      wk.register(opts.defaults)
+    end,
   },
 
   -- Core Utils
@@ -31,120 +209,160 @@ return {
     config = function() require("config.editor.whichkey") end,
   },
 
+  -- git signs highlights text that has changed since the list
+  -- git commit, and also lets you interactively stage & unstage
+  -- hunks in a commit.
   {
     "lewis6991/gitsigns.nvim",
-    event = { "BufReadPre", "BufNewFile" },
-    config = function() require("config.editor.gitsigns") end,
-  },
-
-  {
-    "RRethy/vim-illuminate",
-    event = { "BufReadPost", "BufNewFile" },
-    config = function() require("config.editor.vim-illuminate") end,
-  },
-
-  -- Project Management
-  -- {
-  --   "Shatur/neovim-session-manager",
-  --   config = function() require("config.editor.neovim-session-manager") end,
-  -- },
-
-  {
-    "coffebar/neovim-project",
-    dependencies = {
-      { "nvim-lua/plenary.nvim" },
-      { "nvim-telescope/telescope.nvim" },
-      { "Shatur/neovim-session-manager" },
-    },
-    lazy = false,
-    priority = 100,
-    init = function()
-      -- enable saving the state of plugins in the session
-      vim.opt.sessionoptions:append("globals") -- save global variables that start with an uppercase letter and contain at least one lowercase letter.
-    end,
-    keys = require("config.editor.neovim-project").keys,
-    opts = require("config.editor.neovim-project").opts,
-  },
-
-  -- Buffer Folding
-
-  {
-    "kevinhwang91/nvim-ufo",
-    event = { "BufReadPost", "BufNewFile" },
-    dependencies = { "kevinhwang91/promise-async", event = "BufReadPost" },
+    event = "VeryLazy",
     opts = {
-      fold_virt_text_handler = function(virtText, lnum, endLnum, width, truncate)
-        local newVirtText = {}
-        local suffix = ("  … %d "):format(endLnum - lnum)
-        local sufWidth = vim.fn.strdisplaywidth(suffix)
-        local targetWidth = width - sufWidth
-        local curWidth = 0
-        for _, chunk in ipairs(virtText) do
-          local chunkText = chunk[1]
-          local chunkWidth = vim.fn.strdisplaywidth(chunkText)
-          if targetWidth > curWidth + chunkWidth then
-            table.insert(newVirtText, chunk)
+      signs = {
+        add = { text = "▎" },
+        change = { text = "▎" },
+        delete = { text = "" },
+        topdelete = { text = "" },
+        changedelete = { text = "▎" },
+        untracked = { text = "▎" },
+      },
+      -- signs_staged = {
+      --   add = { text = "▎" },
+      --   change = { text = "▎" },
+      --   delete = { text = "" },
+      --   topdelete = { text = "" },
+      --   changedelete = { text = "▎" },
+      -- },
+      on_attach = function(buffer)
+        local gs = package.loaded.gitsigns
+
+        local function map(mode, l, r, desc) vim.keymap.set(mode, l, r, { buffer = buffer, desc = desc }) end
+
+        -- stylua: ignore start
+        map("n", "]h", function()
+          if vim.wo.diff then
+            vim.cmd.normal({ "]c", bang = true })
           else
-            chunkText = truncate(chunkText, targetWidth - curWidth)
-            local hlGroup = chunk[2]
-            table.insert(newVirtText, { chunkText, hlGroup })
-            chunkWidth = vim.fn.strdisplaywidth(chunkText)
-            -- str width returned from truncate() may less than 2nd argument, need padding
-            if curWidth + chunkWidth < targetWidth then
-              suffix = suffix .. (" "):rep(targetWidth - curWidth - chunkWidth)
-            end
-            break
+            gs.nav_hunk("next")
           end
-          curWidth = curWidth + chunkWidth
-        end
-        table.insert(newVirtText, { suffix, "MoreMsg" })
-        return newVirtText
+        end, "Next Hunk")
+        map("n", "[h", function()
+          if vim.wo.diff then
+            vim.cmd.normal({ "[c", bang = true })
+          else
+            gs.nav_hunk("prev")
+          end
+        end, "Prev Hunk")
+        map("n", "]H", function() gs.nav_hunk("last") end, "Last Hunk")
+        map("n", "[H", function() gs.nav_hunk("first") end, "First Hunk")
+        map({ "n", "v" }, "<leader>ghs", ":Gitsigns stage_hunk<CR>", "Stage Hunk")
+        map({ "n", "v" }, "<leader>ghr", ":Gitsigns reset_hunk<CR>", "Reset Hunk")
+        map("n", "<leader>ghS", gs.stage_buffer, "Stage Buffer")
+        map("n", "<leader>ghu", gs.undo_stage_hunk, "Undo Stage Hunk")
+        map("n", "<leader>ghR", gs.reset_buffer, "Reset Buffer")
+        map("n", "<leader>ghp", gs.preview_hunk_inline, "Preview Hunk Inline")
+        map("n", "<leader>ghb", function() gs.blame_line({ full = true }) end, "Blame Line")
+        map("n", "<leader>ghB", function() gs.blame() end, "Blame Buffer")
+        map("n", "<leader>ghd", gs.diffthis, "Diff This")
+        map("n", "<leader>ghD", function() gs.diffthis("~") end, "Diff This ~")
+        map({ "o", "x" }, "ih", ":<C-U>Gitsigns select_hunk<CR>", "GitSigns Select Hunk")
       end,
-      open_fold_hl_timeout = 0,
     },
+  },
+
+  -- better diagnostics list and others
+  {
+    "folke/trouble.nvim",
+    cmd = { "Trouble" },
+    opts = {},
     keys = {
-      { "fd", "zd", desc = "Delete fold under cursor" },
-      { "fo", "zo", desc = "Open fold under cursor" },
-      { "fO", "zO", desc = "Open all folds under cursor" },
-      { "fc", "zC", desc = "Close all folds under cursor" },
-      { "fa", "za", desc = "Toggle fold under cursor" },
-      { "fA", "zA", desc = "Toggle all folds under cursor" },
-      { "fv", "zv", desc = "Show cursor line" },
+      { "<leader>xx", "<cmd>Trouble diagnostics toggle<cr>",              desc = "Diagnostics (Trouble)" },
+      { "<leader>xX", "<cmd>Trouble diagnostics toggle filter.buf=0<cr>", desc = "Buffer Diagnostics (Trouble)" },
+      { "<leader>cs", "<cmd>Trouble symbols toggle focus=false<cr>",      desc = "Symbols (Trouble)" },
       {
-        "fM",
-        function() require("ufo").closeAllFolds() end,
-        desc = "Close all folds",
+        "<leader>cS",
+        "<cmd>Trouble lsp toggle focus=false win.position=right<cr>",
+        desc = "LSP references/definitions/... (Trouble)",
+      },
+      { "<leader>xL", "<cmd>Trouble loclist toggle<cr>", desc = "Location List (Trouble)" },
+      { "<leader>xQ", "<cmd>Trouble qflist toggle<cr>",  desc = "Quickfix List (Trouble)" },
+      {
+        "[q",
+        function()
+          if require("trouble").is_open() then
+            require("trouble").prev({ skip_groups = true, jump = true })
+          else
+            local ok, err = pcall(vim.cmd.cprev)
+            if not ok then vim.notify(err, vim.log.levels.ERROR) end
+          end
+        end,
+        desc = "Previous Trouble/Quickfix Item",
       },
       {
-        "fR",
-        function() require("ufo").openAllFolds() end,
-        desc = "Open all folds",
+        "]q",
+        function()
+          if require("trouble").is_open() then
+            require("trouble").next({ skip_groups = true, jump = true })
+          else
+            local ok, err = pcall(vim.cmd.cnext)
+            if not ok then vim.notify(err, vim.log.levels.ERROR) end
+          end
+        end,
+        desc = "Next Trouble/Quickfix Item",
       },
-      { "fm", "zm", desc = "Fold more" },
-      { "fr", "zr", desc = "Fold less" },
-      { "fx", "zx", desc = "Update folds" },
-      { "fz", "zz", desc = "Center this line" },
-      { "ft", "zt", desc = "Top this line" },
-      { "fb", "zb", desc = "Bottom this line" },
-      { "fg", "zg", desc = "Add word to spell list" },
-      { "fw", "zw", desc = "Mark word as bad/misspelling" },
-      { "fe", "ze", desc = "Right this line" },
-      { "fE", "zE", desc = "Delete all folds in current buffer" },
-      { "fs", "zs", desc = "Left this line" },
-      { "fH", "zH", desc = "Half screen to the left" },
-      { "fL", "zL", desc = "Half screen to the right" },
     },
   },
 
+  -- Finds and lists all of the TODO, HACK, BUG, etc comment
+  -- in your project and loads them into a browsable list.
   {
-    "luukvbaal/statuscol.nvim",
-    event = { "BufReadPost", "BufNewFile" },
-    config = function() require("config.editor.statuscol") end,
+    "folke/todo-comments.nvim",
+    cmd = { "TodoTrouble", "TodoTelescope" },
+    event = "VeryLazy",
+    opts = {},
+    -- stylua: ignore
+    keys = {
+      { "]t",         function() require("todo-comments").jump_next() end,              desc = "Next Todo Comment" },
+      { "[t",         function() require("todo-comments").jump_prev() end,              desc = "Previous Todo Comment" },
+      { "<leader>xt", "<cmd>Trouble todo toggle<cr>",                                   desc = "Todo (Trouble)" },
+      { "<leader>xT", "<cmd>Trouble todo toggle filter = {tag = {TODO,FIX,FIXME}}<cr>", desc = "Todo/Fix/Fixme (Trouble)" },
+      { "<leader>st", "<cmd>TodoTelescope<cr>",                                         desc = "Todo" },
+      { "<leader>sT", "<cmd>TodoTelescope keywords=TODO,FIX,FIXME<cr>",                 desc = "Todo/Fix/Fixme" },
+    },
   },
 
-  -- Fix Last Place Jump
+  -- Terminal
   {
-    "ethanholz/nvim-lastplace",
-    config = function() require("nvim-lastplace").setup({}) end,
+    "akinsho/toggleterm.nvim",
+    event = { "BufReadPost", "BufNewFile" },
+    opts = {
+      open_mapping = [[<C-\>]],
+      start_in_insert = true,
+      autochdir = false,
+      shade_terminals = true,
+      highlights = {
+        FloatBorder = { link = "ToggleTermBorder" },
+        Normal = { link = "ToggleTerm" },
+        NormalFloat = { link = "ToggleTerm" },
+      },
+      winbar = {
+        enabled = false,
+        name_formatter = function(term) return term.name end,
+      },
+    },
+    config = function(_, opts)
+      local toggleterm_group = vim.api.nvim_create_augroup("UserToggleTerm", { clear = true })
+      vim.api.nvim_create_autocmd("TermOpen", {
+        pattern = "term://*",
+        group = toggleterm_group,
+        callback = function()
+          local options = { buffer = 0 }
+          vim.keymap.set("t", "<esc>", [[<C-\><C-n>]], options)
+          vim.keymap.set("t", "jk", [[<C-\><C-n>]], options)
+          vim.keymap.set("t", "<C-g>", [[<C-\><C-n>]], options)
+        end,
+      })
+
+      require("toggleterm").setup(opts)
+    end,
   },
+
 }
