+++ lua/plugins/extras/coding/icon-picker.lua
@@ -1,0 +1,10 @@
+return {
+  {
+    "ziontee113/icon-picker.nvim",
+    config = function()
+      require("icon-picker").setup({
+        disable_legacy_commands = true,
+      })
+    end,
+  },
+}
