+++ lua/plugins/extras/coding/neogen.lua
@@ -11,22 +11,23 @@ return {
     },
   },
   opts = function(_, opts)
-    if opts.snippet_engine ~= nil then
-      return
-    end
-
-    local map = {
-      ["LuaSnip"] = "luasnip",
-      ["nvim-snippy"] = "snippy",
-      ["vim-vsnip"] = "vsnip",
-    }
-
-    for plugin, engine in pairs(map) do
-      if CoreUtil.has(plugin) then
-        opts.snippet_engine = engine
-        return
-      end
-    end
+    -- TODO debug why luasnip is messing with buffer highlighting
+    -- if opts.snippet_engine ~= nil then
+    --   return
+    -- end
+    --
+    -- local map = {
+    --   ["LuaSnip"] = "luasnip",
+    --   ["nvim-snippy"] = "snippy",
+    --   ["vim-vsnip"] = "vsnip",
+    -- }
+    --
+    -- for plugin, engine in pairs(map) do
+    --   if CoreUtil.has(plugin) then
+    --     opts.snippet_engine = engine
+    --     return
+    --   end
+    -- end
 
     if vim.snippet then
       opts.snippet_engine = "nvim"
