+++ lua/plugins/extras/editor/markdown.lua
@@ -1,0 +1,6 @@
+return {
+  {
+    "iamcco/markdown-preview.nvim",
+    build = "cd app && yarn install",
+  },
+}
